requirements.txt
Pillow==2.5.1
sqlalchemy==0.9.8
-e ./
